# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "description": "API for Social Media application",
        "title": "Social Media Application API"
    },
    "host": "localhost:3000",
    "securityDefinitions": {
        "jwt": {}
    },
    "paths": {
...
        "/api/comments/{postId}": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "To fetch logged-in user comments for mentioned postId.",
                "description": "Only logged-in user comments for mentioned postId will fetch.",
                "operationId": "getAllCommentsById",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "description": "The id of the post to get all comments on it.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "In result key of response object you will find all comments of respective postId and userId."
                    },
                    "404": {
                        "description": "Respective post comments not found. Please verify postId that you want."
                    }
                }
            },
            "post": {
                "tags": [
                    "Comments"
                ],
                "summary": "Comment on specific post.",
                "description": "Only authenticated user will comment on mentioned Post.",
                "operationId": "createCommentById",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "description": "The id of the post to which the comment is.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Content of the comment.",
                        "properties": {
                            "content": {
                                "type": "string",
                                "required": true
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "In result key of response object you will find newly added comment and message as comment added successfully."
                    },
                    "400": {
                        "description": "Error while uploading comment. Please verify postId that you want."
                    }
                }
            }
        },
        "/api/comments/{commentId}": {
            "put": {
                "tags": [
                    "Comments"
                ],
                "summary": "Updates content of a comment.",
                "description": "Only authenticated user can update a mentioned id comment.",
                "operationId": "updateCommentById",
                "parameters": [
                    {
                        "name": "commentId",
                        "in": "path",
                        "description": "The id of the comment to update .",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Content of the comment.",
                        "properties": {
                            "content": {
                                "type": "string",
                                "required": true
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Comment updated successfully."
                    },
                    "400": {
                        "description": "Error while uploading comment. Please verify commentId that you want to update."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Comments"
                ],
                "summary": "Deletes a comment",
                "description": "Only authenticated user deletes a mentioned id comment.",
                "operationId": "deleteCommentById",
                "parameters": [
                    {
                        "name": "commentId",
                        "in": "path",
                        "description": "The id of the comment to delete .",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Comment deleted successfully."
                    },
                    "400": {
                        "description": "Error while deleting comment. Please verify commentId that you want to update."
                    }
                }
            }
        },
        "/api/comments/all/{postId}": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "Get all comments of a post",
                "description": "Fetching all comments of respective postId and display according to pagination parameters.",
                "operationId": "getAllComments",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "description": "The id of the post to fetch all comments on it.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "type": "integer",
                        "description": "",
                        "required": true
                    },
                    {
                        "name": "pageSize",
                        "in": "query",
                        "type": "integer",
                        "description": "",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "In result key of response object you will find all comments of respective postId."
                    },
                    "400": {
                        "description": "Error while fetching comments. Please verify postId that you want to get."
                    }
                }
            }
        },
...
        "/api/bookmarks/": {
            "get": {
                "tags": [
                    "Bookmarks"
                ],
                "summary": "All bookmarked post of a user.",
                "description": "All bookmarked posts of logged-in user.",
                "operationId": "getBookmarksByUser",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "In result key of response object you will find all bookmarks of respective userId."
                    },
                    "404": {
                        "description": "Not yet bookmarked posts."
                    },
                    "400": {
                        "description": "Error occured while bookmark a post."
                    }
                }
            }
        },
        "/api/bookmarks/toggle/{postId}": {
            "get": {
                "tags": [
                    "Bookmarks"
                ],
                "summary": "Bookmark or Remove bookmark a post",
                "description": "Only authenticated user Toggles bookmark or remove bookmark a post with mentioned postId",
                "operationId": "togglePostBookmark",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "description": "The id of the post to fetch all comments on it.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Removed bookmark successfully."
                    },
                    "201": {
                        "description": "Bookmarked successfully."
                    },
                    "400": {
                        "description": "Error occured while bookmark/remove bookmark a post."
                    }
                }
            }
        },
        "/api/friends/get-friends": {
            "get": {
                "tags": [
                    "Friends"
                ],
                "summary": "All friends of a user.",
                "description": "All friends of logged-in user.[ those friends whose friendship status is accepted.]",
                "operationId": "getMyFriends",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "In result key of response object you will find all friends of respective userId."
                    },
                    "404": {
                        "description": "Sorry! You don't have friends yet."
                    }
                }
            }
        },
        "/api/friends/get-pending-requests": {
            "get": {
                "tags": [
                    "Friends"
                ],
                "summary": "To get all pending friend requests.",
                "description": "Pending friend requests of logged in user.[status is pending including all incoming & ougoing friend requests.]",
                "operationId": "getPendingFriendRequests",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "In result key of response object you will find all pending friend requests of respective userId."
                    },
                    "404": {
                        "description": "No pending request."
                    }
                }
            }
        },
        "/api/friends/toggle-friendship/{friendId}": {
            "get": {
                "tags": [
                    "Friends"
                ],
                "summary": "Easy mechanism to accept or reject friendship.",
                "description": "Accept or recject friendship with mentioned friend.",
                "operationId": "toggleFriendship",
                "parameters": [
                    {
                        "name": "friendId",
                        "in": "path",
                        "description": "A friend ID to do friendship or reject it.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Oops! Your friendship broken."
                    },
                    "201": {
                        "description": "Yeah! you are now friends of each other."
                    },
                    "400": {
                        "description": "Error occured or You are alredy in friendship."
                    }
                }
            }
        },
        "/api/friends/response-to-request/{friendId}/{action}": {
            "get": {
                "tags": [
                    "Friends"
                ],
                "summary": "Response to friend request.",
                "description": "Eigther accept or reject a friend request. [ for all friend request received to user and are in pending state. ]",
                "operationId": "friendshipAcceptReject",
                "parameters": [
                    {
                        "name": "friendId",
                        "in": "path",
                        "description": "A friend ID to do friendship or reject it.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "action",
                        "in": "path",
                        "description": "Action refers to accept, reject or pending.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Friendship accepted. or Friendship rejected."
                    },
                    "403": {
                        "description": "Internal server error. Please try after some time."
                    }
                }
            }
        },
        "/api/friends/send-friend-request/{friendId}": {
            "get": {
                "tags": [
                    "Friends"
                ],
                "summary": "Send a friend request.",
                "description": "Send a friend request.",
                "operationId": "sendFriendRequest",
                "parameters": [
                    {
                        "name": "friendId",
                        "in": "path",
                        "description": "A friend ID to send friendship request.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Friend request sent. or You are alredy in friendship."
                    },
                    "500": {
                        "description": "Internal Server Error."
                    }
                }
            }
        },
...
    },
...
}